Remove redundant routine tests
